# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Cloud SaaS POS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: branchId
          in: query
          schema:
            type: ["integer", "null"]
        - name: active
          in: query
          schema:
            type: ["boolean", "null"]
      responses:
        "200":
          description: Products list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─────── INVENTORY ───────
...
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: List purchase orders
      parameters:
        - name: supplierId
          in: query
          schema:
            type: ["integer", "null"]
        - name: status
          in: query
          schema:
            type: ["string", "null"]
        - name: startDate
          in: query
          schema:
            type: ["string", "null"]
        - name: endDate
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Purchases list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Purchase"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Create a purchase order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}:
    get:
      operationId: getPurchase
      tags: [purchases]
      summary: Get purchase detail with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Purchase detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseDetail"
    patch:
      operationId: updatePurchase
      tags: [purchases]
      summary: Update a purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}/receive:
    post:
      operationId: receivePurchase
      tags: [purchases]
      summary: Mark purchase as received and update inventory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
  /purchases/{id}/payments:
    post:
      operationId: addPurchasePayment
      tags: [purchases]
      summary: Record a payment against a purchase
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInput"
      responses:
        "201":
          description: Payment recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"

  # ─────── SALES / POS ───────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, name, sellingPrice, costPrice, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        sku:
          type: ["string", "null"]
        barcode:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        brand:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        costPrice:
          type: number
        sellingPrice:
          type: number
        taxRate:
          type: number
        reorderLevel:
          type: number
        description:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        currentStock:
          type: ["number", "null"]
        createdAt:
          type: string
        updatedAt:
          type: ["string", "null"]
    ProductInput:
      type: object
      required: [name, sellingPrice, costPrice]
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        categoryId:
          type: ["integer", "null"]
        brand:
          type: string
        unit:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        taxRate:
          type: number
        reorderLevel:
          type: number
        description:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        categoryId:
          type: ["integer", "null"]
        brand:
          type: string
        unit:
          type: string
        costPrice:
          type: number
        sellingPrice:
          type: number
        taxRate:
          type: number
        reorderLevel:
          type: number
        description:
          type: string
        imageUrl:
          type: string
        isActive:
          type: boolean

    # ─────── INVENTORY ───────
...
    Purchase:
      type: object
      required: [id, supplierId, supplierName, status, totalAmount, paidAmount, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        branchId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, received, partial, paid]
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceDue:
          type: number
        invoiceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PurchaseDetail:
      type: object
      required: [id, supplierId, supplierName, status, totalAmount, paidAmount, items, payments, createdAt]
      properties:
        id:
          type: integer
        supplierId:
          type: integer
        supplierName:
          type: string
        branchId:
          type: ["integer", "null"]
        status:
          type: string
        totalAmount:
          type: number
        paidAmount:
          type: number
        balanceDue:
          type: number
        invoiceNumber:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItem"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
        createdAt:
          type: string
    PurchaseItem:
      type: object
      required: [id, productId, productName, quantity, unitCost, total]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: number
        unitCost:
          type: number
        total:
          type: number
    PurchaseInput:
      type: object
      required: [supplierId, items]
      properties:
        supplierId:
          type: integer
        branchId:
          type: ["integer", "null"]
        invoiceNumber:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseItemInput"
    PurchaseItemInput:
      type: object
      required: [productId, quantity, unitCost]
      properties:
        productId:
          type: integer
        quantity:
          type: number
        unitCost:
          type: number
    PurchaseUpdate:
      type: object
      properties:
        invoiceNumber:
          type: string
        notes:
          type: string
        status:
          type: string

    # ─────── PAYMENT ───────
    Payment:
      type: object
      required: [id, amount, method, createdAt]
      properties:
        id:
          type: integer
        amount:
          type: number
        method:
          type: string
          enum: [cash, mobile_money, card, bank_transfer, credit]
        reference:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PaymentInput:
      type: object
      required: [amount, method]
      properties:
        amount:
          type: number
        method:
          type: string
          enum: [cash, mobile_money, card, bank_transfer, credit]
        reference:
          type: string
        notes:
          type: string

    # ─────── SALE ───────
...